# Optimize InstanceNormOp forward